refactor: 個人成績表のプレビューとPDF出力を統合リファクタリング
プレビュー（React）とPDF出力（HTML文字列テンプレート）が完全に別の
レンダリングパスで実装されていた重複を解消。renderToStaticMarkupを使い、
プレビュー用Reactコンポーネントを再利用してPDF用HTMLを生成する方式に統合。

主な変更:
- computeReportData.ts: 共通計算ロジックを純粋関数として集約
- ReportSectionViews.tsx: hooks不使用の純粋ビューコンポーネントを作成
- BoxPlotChart.tsx: BoxPlotChartView（純粋ビュー）を分離エクスポート
- SubtotalTablePreview.tsx: SubtotalTableView（純粋ビュー）を分離エクスポート
- IndividualReportPreview.tsx: 共通関数+純粋ビューを使用するよう修正
- generatePrintHtml.ts: renderToStaticMarkupで全面書き換え（1443行→348行）

全体で約920行の重複コードを削減。
